==> globalprep/fis/v2022/STEP2b_species_resilience_lookup_table.Rmd <==
---
title: "OHI 2022: FP, SAUP Species Resilience Lookup Table"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
  toc: true
number_sections: true
theme: cerulean
highlight: haddock
includes: 
  in_header: '../../../workflow/templates/ohi_hdr.html'
pdf_document:
  toc: true
---

This script uses FishBase to find the Resilience for each of the species in the SAUP database. The Resilience information is needed for running catch-MSY to estimate B/Bmsy.

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE, eval=FALSE)

library(dplyr)
library(rfishbase)
library(parallel)
source('../../../workflow/R/common.R')

```


# Read in SAUP catch data and match common names to scientific names

```{r catch_dat, eval = FALSE}

taxon <- read.csv('output/stock_catch_no_res.csv') %>%
          dplyr::select(sciname = TaxonName,
                 species = CommonName,
                 TaxonKey) %>%
            filter(TaxonKey > 599999) %>%
            distinct()

# validate scientific names with fishbase

sp <- taxon$sciname

fb <- validate_names(sp)

resil <- mclapply(fb, rfishbase::stocks, fields = c("Resilience", "Species", "SpecCode"), mc.cores=1) # more than one core was failing

res <- do.call(rbind.data.frame, resil) %>% 
  distinct() 

res <- res %>%
  rename(sciname = Species)



taxon_res <- taxon %>%
            left_join(res, by='sciname') %>%
            dplyr::select(sciname, common = species, Resilience, SpecCode)


write.csv(taxon_res, "int/taxon_res_int.csv", row.names = FALSE)

taxon_res <- read.csv("int/taxon_res_int.csv")
```

There are some duplicated species since fishbase provides resilience per stock...we'll want to remove the duplicates 

```{r duplicates, eval = FALSE}
dup <- taxon_res$sciname[duplicated(taxon_res$sciname)]

d <- taxon_res %>%
      filter(sciname %in% dup)%>% #select those species that are duplicated
      filter(!is.na(Resilience)) #remove rows where one of those species has an NA for Resilience (this is most species)

dup_sp <- d$sciname[duplicated(d$sciname)]
```

Now we have 2 species that have stocks with 2 different resiliences:
  "Clupea harengus"    "Platichthys flesus"     

For these three species, we will remove all resilience's except for Medium since Medium is the default resilience when there isn't one for catch-MSY. 
```{r, eval = FALSE}

three_sp <- d %>%
            filter(sciname %in% dup_sp,
                   Resilience == "Medium")

#all other duplicated species with the NA resilience's removed
other_sp <- d %>%
            filter(!sciname %in% dup_sp)

taxon_res2 <- taxon_res %>%
              filter(!sciname %in% dup) %>%
              rbind(other_sp) %>%
              rbind(three_sp)
```

We still have species with NA for Resilience. This is likely for non-fish species. We can try to get resilience for these species by setting the URL to SeaLifeBase in the stocks call.

```{r, eval = FALSE}

sp_w_nas <- taxon_res2 %>%
            filter(is.na(Resilience)) %>%
            .$sciname

write.csv(sp_w_nas, "int/sp_w_nas.csv", row.names = FALSE)

resil <- mclapply(sp_w_nas, stocks, fields = c("Resilience","Species"), mc.cores=1, server = "sealifebase")

# test <- stocks(sp_w_nas[2], fields = c("Resilience", "Species"), server = "sealifebase")

res <- do.call(rbind.data.frame, resil)%>%
        distinct()

DT::datatable(res)

res_info_na <- res %>%
  filter(!is.na(Resilience)) ## These are the species which have resilience info in SeaLifeBase, but not in our taxon_res2 dataframe. Join these with that df and wrangle to fill in these missing data. 

taxon_res3 <- taxon_res2 %>%
  left_join(res_info_na, by = c("sciname" = "Species")) %>%
  mutate(Resilience = ifelse(is.na(Resilience.x), Resilience.y, Resilience.x)) %>%
  dplyr::select(-Resilience.x, -Resilience.y)
```

We will have to leave the rest as NA (there aren't very many). This will cause the CMSY model to default to Medium resilience but will not remove them from assessment.

```{r, eval = FALSE}
write.csv(taxon_res3,file = 'output/taxon_resilience_lookup.csv')
```
